Task 2 from the fifth block of tasks for working with a .csv file
The user is asked how many records he wants to add to the list, then
the user enters the records. After the data is added,
the author is requested from the user and all the books of the specified author are displayed from the list. If
there are no books by this author in the list, print the corresponding message. To solve the problem, the books.csv file is used