Task 4 from the first block
I wrote a program that solves the following problem: In the soccer championship the team is given 3 points for winning,  0 points for losing, and 0 points for a draw. - 1. The number of points received by the team for the game is known. Determined the verbal the result of the game (win, lose or draw).  The number is entered from the console.